// src/app/services/auth/models/tenant-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TenantDto {
  createdAt?: string;
  currentUsers?: number;
  id?: string;
  isActive?: boolean;
  maxUsers?: number;
  name?: string;
  planType?: 'FREE' | 'BASIC' | 'PREMIUM' | 'ENTERPRISE';
  subdomain?: string;
}